refactor: move app to home component
